update Auto controller

+++ src/main/java/ru/ekorzunov/urfu_bach_prog_3_coursework/controller/AutoController.java
@@ -1,15 +1,19 @@
 package ru.ekorzunov.urfu_bach_prog_3_coursework.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.security.access.AccessDeniedException;
 import ru.ekorzunov.urfu_bach_prog_3_coursework.entity.Auto;
 import ru.ekorzunov.urfu_bach_prog_3_coursework.entity.User;
 import ru.ekorzunov.urfu_bach_prog_3_coursework.repository.AutoRepository;
 import ru.ekorzunov.urfu_bach_prog_3_coursework.repository.UserRepository;
+import ru.ekorzunov.urfu_bach_prog_3_coursework.service.UserDetailsServiceImpl;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,6 +28,14 @@ public class AutoController {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private UserDetailsServiceImpl userDetailsService;
+
+    @GetMapping("/")
+    public String getAutos() {
+        return "redirect:./list";
+    }
 
     @GetMapping("/list")
     public ModelAndView getAllAutos() {
@@ -42,35 +54,68 @@ public class AutoController {
     }
 
     @GetMapping("/add")
-    public ModelAndView addAutoForm() {
+    public ModelAndView addAutoForm(@AuthenticationPrincipal UserDetails userDetail, HttpServletRequest request) {
         ModelAndView mav = new ModelAndView("add-auto-form");
-        Auto auto = new Auto();
-        mav.addObject("auto", auto);
+
+        List<User> users;
+        boolean is_admin = request.isUserInRole("ROLE_ADMIN");
+        if (is_admin) {
+            users = userRepository.findAll();
+        } else {
+            User currentUser = userRepository.findByEmail(userDetail.getUsername());
+            users = List.of(currentUser);
+        }
+
+        mav.addObject("auto", new Auto());
+        mav.addObject("users", users);
         return mav;
     }
 
     @PostMapping("/save")
-    public String saveAuto(@ModelAttribute Auto auto, @AuthenticationPrincipal User user) {
-        auto.setOwner(user);
+    public String saveAuto(@ModelAttribute Auto auto, @AuthenticationPrincipal UserDetails userDetail, HttpServletRequest request) {
+        boolean is_admin = request.isUserInRole("ROLE_ADMIN");
+        if (!is_admin) {
+            User user = userRepository.findByEmail(userDetail.getUsername());
+            auto.setOwner(user);
+        }
         autoRepository.save(auto);
         return "redirect:./list";
     }
 
     @GetMapping("/update")
-    public ModelAndView showUpdateForm(@RequestParam Long autoId) {
+    public ModelAndView showUpdateForm(@RequestParam Long autoId, @AuthenticationPrincipal UserDetails userDetail, HttpServletRequest request) {
         ModelAndView mav = new ModelAndView("add-auto-form");
+
         Optional<Auto> optionalAuto = autoRepository.findById(autoId);
         Auto auto = new Auto();
         if (optionalAuto.isPresent()) {
             auto = optionalAuto.get();
         }
         mav.addObject("auto", auto);
+
+        List<User> users;
+        boolean is_admin = request.isUserInRole("ROLE_ADMIN");
+        if (is_admin) {
+            users = userRepository.findAll();
+        } else {
+            User user = userRepository.findByEmail(userDetail.getUsername());
+            users = List.of(user);
+        }
+        mav.addObject("users", users);
+
         return mav;
     }
 
     @GetMapping("/delete")
-    public String deleteAuto(@RequestParam Long autoId) {
-        autoRepository.deleteById(autoId);
+    public String deleteAuto(@RequestParam Long autoId, @AuthenticationPrincipal UserDetails userDetail, HttpServletRequest request) {
+        User user = userRepository.findByEmail(userDetail.getUsername());
+        boolean is_admin = request.isUserInRole("ROLE_ADMIN");
+        boolean is_owner = autoRepository.findById(autoId).orElseThrow().getOwner().equals(user);
+        if (is_admin || is_owner) {
+            autoRepository.deleteById(autoId);
+        } else {
+            throw new AccessDeniedException("403 returned");
+        }
         return "redirect:./list";
     }
